requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
multidict==6.6.3
mypy==1.17.0
mypy_extensions==1.1.0
pathspec==0.12.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
ruff==0.12.4
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
